+++ packages/codegen/src/cli.ts
@@ -1,20 +1,18 @@
 #! /usr/bin/env node
+import { getConfig } from "@powerhousedao/config/powerhouse";
 import {
   generate,
   generateEditor,
   generateFromFile,
   generateProcessor,
+  generateSubgraph,
 } from "./codegen/index";
-import {
-  parseArgs,
-  getConfig,
-  promptDirectories,
-  parseConfig,
-} from "./utils/index";
+import { parseArgs, promptDirectories, parseConfig } from "./utils/index";
 
 function parseCommand(argv: string[]) {
   const args = parseArgs(argv, {
     "--editor": String,
+    "--subgraph": String,
     "-e": "--editor",
     "--processor": String,
     "--document-types": String,
@@ -23,6 +21,7 @@ function parseCommand(argv: string[]) {
   const editorName = args["--editor"];
   const processorName = args["--processor"];
   const processorType = args["--processor-type"];
+  const subgraphName = args["--subgraph"];
   return {
     processor: !!processorName,
     processorName,
@@ -31,6 +30,8 @@ function parseCommand(argv: string[]) {
     editorName,
     documentTypes: args["--document-types"],
     arg: args._,
+    subgraph: !!subgraphName,
+    subgraphName: subgraphName ?? "example",
   };
 }
 
@@ -60,8 +61,9 @@ async function main() {
       throw new Error("processor name is required (--processor)");
     }
 
-    const type =
-      command.processorType === "analytics" ? "analytics" : "operational";
+    const type = !command.processorType
+      ? "analytics"
+      : (command.processorType as "analytics" | "operational");
 
     await generateProcessor(
       command.processorName,
@@ -69,6 +71,8 @@ async function main() {
       command.documentTypes?.split(/[|,;]/g) ?? [],
       config,
     );
+  } else if (command.subgraph) {
+    await generateSubgraph(command.subgraphName, config);
   } else if (command.arg.length === 2) {
     await generateFromFile(command.arg[1], config);
   } else {