+++ packages/codegen/src/codegen/.hygen/templates/powerhouse/generate-processor-analytics/lib.inject_export.esm.t
@@ -4,4 +4,4 @@ append: true
 to: "<%= rootDir %>/index.ts"
 skip_if: "<%= h.changeCase.pascal(name) %>"
 ---
-export * as <%= h.changeCase.pascal(name) %>AnalyticsProcessor from "./<%= h.changeCase.param(name)  %>/src";
+export * as <%= h.changeCase.pascal(name) %>Processor from "./<%= h.changeCase.param(name)  %>";