+++ packages/document-drive/vitest.config.ts
@@ -1,8 +1,10 @@
+import { loadEnv } from "vite";
 import { defineConfig } from "vitest/config";
 
-export default defineConfig({
+export default defineConfig(({ mode }) => ({
   test: {
     environment: "node",
+    env: loadEnv(mode, process.cwd(), ""),
     testTimeout: 5000,
     server: {
       deps: {
@@ -16,4 +18,4 @@ export default defineConfig({
       },
     },
   },
-});
+}));