+++ packages/document-model/src/document/utils/index.ts
@@ -16,6 +16,7 @@ export {
   sortOperations,
 } from "./base";
 export * from "./crypto";
+export * from "./errors";
 export * from "./file";
 export * from "./validation";
 export * as documentHelpers from "./document-helpers";