+++ packages/reactor-api/src/server.ts
@@ -1,13 +1,20 @@
 import { PGlite } from "@electric-sql/pglite";
 import { IDocumentDriveServer } from "document-drive";
 import express, { Express } from "express";
-import pg from "pg";
-const { Pool } = pg;
-import { ReactorRouterManager } from "./router";
+import { Pool } from "pg";
+import { ProcessorManager } from "./processors";
+import { SubgraphManager } from "./subgraphs/manager";
+import { API } from "./types";
+import { getDbClient } from "./utils/db";
+import {
+  KnexAnalyticsStore,
+  KnexQueryExecutor,
+} from "@powerhousedao/analytics-engine-knex";
 
 type Options = {
   express?: Express;
   port?: number;
+  dbPath: string | undefined;
   client?: PGlite | typeof Pool | undefined;
 };
 
@@ -15,20 +22,25 @@ const DEFAULT_PORT = 4000;
 
 export async function startAPI(
   reactor: IDocumentDriveServer,
-  options: Options
-) {
+  options: Options,
+): Promise<API> {
   const port = options.port ?? DEFAULT_PORT;
   const app = options.express ?? express();
-
-  const reactorRouterManager = new ReactorRouterManager(
+  const db = getDbClient(options.dbPath);
+  const analyticsStore = new KnexAnalyticsStore({
+    executor: new KnexQueryExecutor(),
+    knex: db,
+  });
+  const subgraphManager = new SubgraphManager(
     "/",
     app,
     reactor,
-    options.client
+    db,
+    analyticsStore,
   );
-  await reactorRouterManager.init();
+  await subgraphManager.init();
+  const processorManager = new ProcessorManager(reactor, db, analyticsStore);
 
   app.listen(port);
-
-  return { app, reactorRouterManager };
+  return { app, subgraphManager, processorManager };
 }