+++ packages/reactor-api/src/types.ts
@@ -1,22 +1,43 @@
-import {
-  IDocumentDriveServer,
-  InternalTransmitterUpdate,
-  Listener,
-} from "document-drive";
-import { PgDatabase } from "drizzle-orm/pg-core";
-import { IncomingHttpHeaders } from "http";
+import { IDocumentDriveServer, IReceiver, Listener } from "document-drive";
+import { Document, OperationScope } from "document-model/document";
+import { Express } from "express";
+import { ProcessorClass } from "./processors/processor";
+import { SubgraphManager } from "./subgraphs/manager";
+import { Db } from "./utils/db";
+import { IAnalyticsStore } from "./processors/analytics-processor";
 
-export interface Context {
-  headers: IncomingHttpHeaders;
-  driveId: string | undefined;
-  driveServer: IDocumentDriveServer;
-  db: PgDatabase<any, any, any>;
-}
+export type { Db } from "./utils/db";
 
-export type Processor = {
-  name: string;
-  resolvers: any;
-  typeDefs: string;
-  options?: Omit<Listener, "driveId">;
-  transmit?: (strands: InternalTransmitterUpdate[]) => Promise<void>;
+export type IProcessorManager = {
+  registerProcessor(module: IProcessor | ProcessorClass): Promise<IProcessor>;
 };
+
+export type API = {
+  app: Express;
+  subgraphManager: SubgraphManager;
+  processorManager: IProcessorManager;
+};
+
+export type ProcessorType = "analytics" | "operational";
+
+export type ProcessorSetupArgs = {
+  reactor: IDocumentDriveServer;
+  operationalStore: Db;
+  analyticsStore: IAnalyticsStore;
+};
+
+export type IProcessor<
+  D extends Document = Document,
+  S extends OperationScope = OperationScope,
+> = IReceiver<D, S> & {
+  onSetup?: (args: ProcessorSetupArgs) => void;
+  getOptions: () => ProcessorOptions;
+};
+
+// export interface ProcessorType<T> extends Function {
+//   new (...args: any[]): T;
+//   TYPE: string;
+//   OPTIONS: ProcessorOptions;
+// }
+
+export type ProcessorOptions = Omit<Listener, "driveId"> & { label: string };